backend/payroll_system/api/routers.py
from rest_framework.routers import SimpleRouter

from api.auth.viewsets import LoginViewSet, RegistrationViewSet, RefreshViewSet
from api.views import UserViewSet


routes = SimpleRouter()

# AUTHENTICATION
routes.register(r'auth/login', LoginViewSet, basename='auth-login')
routes.register(r'auth/register', RegistrationViewSet, basename='auth-register')
routes.register(r'auth/refresh', RefreshViewSet, basename='auth-refresh')

# USER
routes.register(r'user', UserViewSet, basename='user')


urlpatterns = [
    *routes.urls
]